Add logic tests structure, remove old tests
